partial localization, css fixes for edit view

+++ database/migrations/2017_03_18_164514_create_posts_table.php
@@ -23,7 +23,7 @@ class CreatePostsTable extends Migration
             $table->string('category');
             $table->integer('price')->unsigned()->index();
             $table->integer('size')->unsigned()->index();
-            $table->integer('rooms')->unsigned()->index();
+            $table->float('rooms')->unsigned()->index();
             $table->string('empty');
             $table->string('housetype');
             $table->string('heating');
